refactor: 마/상 MoveRule로 위임
